+++ src/resources/collections/get.ts
@@ -14,7 +14,7 @@ export default class Get extends AbstractGet {
     return optparse
       .subCommand<GetOptions, GetArgs>('collections [name]')
       .option('-l, --logs [name]', 'Request resource logs')
-      .option('-t, --trace [name]', 'Request resource logs including stacktraces')
+      .option('-T, --trace [name]', 'Request resource logs including stacktraces')
       .alias('co')
       .description('Get collections')
       .action(async (opts, args, rest) => {